+++ app/src/main/java/com/vanniktech/emoji/sample/MainDialog5.java
@@ -24,6 +24,7 @@ import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
@@ -32,24 +33,25 @@ import androidx.core.content.ContextCompat;
 import androidx.fragment.app.DialogFragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
 import com.vanniktech.emoji.EmojiEditText;
-import com.vanniktech.emoji.EmojiPopup;
+import com.vanniktech.emoji.EmojiPopupGeneral;
 import com.vanniktech.emoji.material.MaterialEmojiLayoutFactory;
 
 // We don't care about duplicated code in the sample.
-public class MainDialog extends DialogFragment {
+public class MainDialog5 extends DialogFragment {
   static final String FRAGMENT_MANAGER_TAG = "dialog_main";
   static final String TAG = "MainDialog";
 
   ChatAdapter chatAdapter;
-  EmojiPopup emojiPopup;
+  EmojiPopupGeneral emojiPopup;
 
   EmojiEditText editText;
   ViewGroup rootView;
   ImageView emojiButton;
 
   public static void show(@NonNull final AppCompatActivity activity) {
-    new MainDialog().show(activity.getSupportFragmentManager(), FRAGMENT_MANAGER_TAG);
+    new MainDialog5().show(activity.getSupportFragmentManager(), FRAGMENT_MANAGER_TAG);
   }
 
   @Override public void onCreate(@Nullable final Bundle savedInstanceState) {
@@ -76,7 +78,9 @@ public class MainDialog extends DialogFragment {
     emojiButton.setColorFilter(ContextCompat.getColor(getContext(), R.color.emoji_icons), PorterDuff.Mode.SRC_IN);
     sendButton.setColorFilter(ContextCompat.getColor(getContext(), R.color.emoji_icons), PorterDuff.Mode.SRC_IN);
 
-    emojiButton.setOnClickListener(ignore -> emojiPopup.toggle());
+    emojiButton.setOnClickListener(ignore ->{
+      emojiPopup.toggle();
+    } );
     sendButton.setOnClickListener(ignore -> {
       final String text = editText.getText().toString().trim();
 
@@ -97,7 +101,8 @@ public class MainDialog extends DialogFragment {
   }
 
   private void setUpEmojiPopup() {
-    emojiPopup = EmojiPopup.Builder.fromRootView(rootView)
+    emojiPopup = EmojiPopupGeneral.Builder.fromRootView(rootView)
+            /*
         .setOnEmojiBackspaceClickListener(ignore -> Log.d(TAG, "Clicked on Backspace"))
         .setOnEmojiClickListener((ignore, ignore2) -> Log.d(TAG, "Clicked on emoji"))
         .setOnEmojiPopupShownListener(() -> emojiButton.setImageResource(R.drawable.ic_keyboard))
@@ -105,7 +110,8 @@ public class MainDialog extends DialogFragment {
         .setOnEmojiPopupDismissListener(() -> emojiButton.setImageResource(R.drawable.emoji_ios_category_smileysandpeople))
         .setOnSoftKeyboardCloseListener(() -> Log.d(TAG, "Closed soft keyboard"))
         .setKeyboardAnimationStyle(R.style.emoji_fade_animation_style)
-        .setPageTransformer(new PageTransformer2())
+
+            */     .setPageTransformer(new PageTransformer2())
         .build(editText);
   }
 }